Create pokemon_type_tester.py
the program starts by defining a dictionary (type_data) that contains each Pokemon type and its corresponding weaknesses. It also defines a list of Pokemon types (type_list) that the user will be tested on.

The ask_question function chooses a random Pokemon type from the type_list, prompts the user to enter the weaknesses of that type, and checks the user's answer against the correct answer. If the user's answer is correct, the function returns True; otherwise, it returns False.

The program then sets up variables to keep track of the user's score and the number of questions asked. It then loops 10 times, calling ask_question for each iteration and updating the user's score accordingly.

Finally, the program displays the user's score and the number of questions asked.